Move wait_for_ongoing_refresh into task.rs
